+++ examples/expo-todomvc-sync-cf/src/Root.tsx
@@ -1,7 +1,7 @@
 import { makePersistedAdapter } from '@livestore/adapter-expo'
 import { nanoid } from '@livestore/livestore'
 import { LiveStoreProvider } from '@livestore/react'
-import { makeCfSync } from '@livestore/sync-cf/client'
+import { makeWsSync } from '@livestore/sync-cf/client'
 import { StatusBar } from 'expo-status-bar'
 import React from 'react'
 import { Button, unstable_batchedUpdates as batchUpdates, StyleSheet, Text, View } from 'react-native'
@@ -16,7 +16,7 @@ const storeId = process.env.EXPO_PUBLIC_LIVESTORE_STORE_ID
 const syncUrl = process.env.EXPO_PUBLIC_LIVESTORE_SYNC_URL
 
 const adapter = makePersistedAdapter({
-  sync: { backend: syncUrl ? makeCfSync({ url: syncUrl }) : undefined },
+  sync: { backend: syncUrl ? makeWsSync({ url: syncUrl }) : undefined },
 })
 
 export const Root = () => {