+++ packages/@livestore/sync-cf/src/cf-worker/mod.ts
@@ -1,4 +1,5 @@
+export type { CfTypes } from '@livestore/common-cf'
 export { CfDeclare } from '@livestore/common-cf/declare'
-export * from './durable-object.ts'
+export * from './do/durable-object.ts'
 export * from './shared.ts'
 export * from './worker.ts'