+++ packages/@livestore/sync-cf/src/common/mod.ts
@@ -8,4 +8,8 @@ export * as SyncMessage from './sync-message-types.ts'
 export const SearchParamsSchema = Schema.Struct({
   storeId: Schema.String,
   payload: Schema.compose(Schema.StringFromUriComponent, Schema.parseJson(Schema.JsonValue)).pipe(Schema.UndefinedOr),
+  // NOTE `do-rpc` is handled differently
+  transport: Schema.Union(Schema.Literal('http'), Schema.Literal('ws')),
 })
+
+export type SearchParams = typeof SearchParamsSchema.Type